Decouple component and connector for SinglePost
