--- src/containers/SinglePost.js
@@ -1,193 +1,0 @@
-import React from 'react'
-import Post from '../components/Post'
-import { prefetch } from 'react-fetcher'
-import { connect } from 'react-redux'
-import { includes } from 'lodash'
-import { get } from 'lodash/fp'
-import {
-  navigate,
-  notify,
-  setMetaTags,
-  fetchComments,
-  fetchCommunity,
-  fetchPost,
-  unfollowPost,
-  saveCurrentCommunityId,
-  findError
-} from '../actions'
-import { FETCH_POST } from '../actions/constants'
-import { ogMetaTags } from '../util'
-import A from '../components/A'
-import PostEditor from '../components/PostEditor'
-import { scrollToComment } from '../util/scrolling'
-import AccessErrorMessage from '../components/AccessErrorMessage'
-import CoverImagePage from '../components/CoverImagePage'
-import EventPost from '../components/EventPost'
-import ProjectPost from '../components/ProjectPost'
-import ProjectActivityCard from '../components/ProjectActivityCard'
-import { getCurrentCommunity } from '../models/community'
-import { denormalizedPost, getComments, getPost, isMessageThread } from '../models/post'
-import { fetch, ConnectedPostList } from './ConnectedPostList'
-const { array, bool, object, string, func } = React.PropTypes
-
-const subject = 'post'
-
-@prefetch(({ store, dispatch, params: { id }, query }) =>
-  dispatch(fetchPost(id))
-  .then(action => {
-    if (action.error) return action
-    if (action.payload.parent_post_id) {
-      return dispatch(fetchPost(action.payload.parent_post_id))
-        .then(() => action)
-    } else {
-      return action
-    }
-  })
-  .then(action => redirect(store, id) || setupPage(store, id, query, action)))
-@connect((state, { params: { id } }) => {
-  const post = getPost(id, state)
-  const parentPost = post ? getPost(post.parent_post_id, state) : null
-  return {
-    post: post ? denormalizedPost(post, state) : null,
-    parentPost: parentPost ? denormalizedPost(parentPost, state) : null,
-    community: getCurrentCommunity(state),
-    comments: getComments(post, state),
-    editing: !!state.postEdits[id],
-    error: findError(state.errors, FETCH_POST, 'posts', id)
-  }
-})
-export default class SinglePost extends React.Component {
-  static propTypes = {
-    post: object,
-    parentPost: object,
-    community: object,
-    editing: bool,
-    error: string,
-    dispatch: func,
-    location: object
-  }
-
-  static contextTypes = {
-    isMobile: bool,
-    currentUser: object
-  }
-
-  static childContextTypes = { community: object }
-
-  getChildContext () {
-    return { community: this.props.community }
-  }
-
-  render () {
-    const { post, parentPost, comments, community, editing, error, location: { query } } = this.props
-    const { currentUser, isMobile } = this.context
-    if (error) return <AccessErrorMessage error={error} />
-    if (!post || !post.user) return <div className='loading'>Loading...</div>
-    const isChild = !!post.parent_post_id
-
-    return <div>
-      {isMobile && isChild && <div id='mobile-top-bar'>
-        <A className='back' to={`/p/${post.parent_post_id}`}>
-          <span className='left-angle-bracket'>&#x3008;</span>
-          Back to project
-        </A>
-      </div>}
-      <CoverImagePage id='single-post' image={get('banner_url', community)}>
-        {editing
-          ? <PostEditor post={post} parentPost={parentPost} expanded />
-          : showPost(post, parentPost, comments)
-        }
-        {post.type === 'project' && <div>
-          {currentUser &&
-            <PostEditor community={community} parentPost={post} />
-          }
-          <ConnectedPostList
-            subject={subject}
-            id={post.id}
-            parentPost={post}
-            query={{...query}}
-            noPostsMessage='There are no project related conversations to show.' />
-        </div>}
-      </CoverImagePage>
-    </div>
-  }
-}
-
-const showPost = (post, parentPost, comments) => {
-  if (parentPost) {
-    return <ProjectActivityCard {...{post, parentPost}} expanded />
-  }
-  switch (post.type) {
-    case 'event':
-      return <EventPost post={post} comments={comments} />
-    case 'project':
-      return <ProjectPost post={post} comments={comments} />
-    default:
-      return <Post {...{post, parentPost}} expanded />
-  }
-}
-
-const redirect = (store, id) => {
-  const state = store.getState()
-  const post = state.posts[id]
-  if (isMessageThread(post)) {
-    store.dispatch(navigate(`/t/${post.id}`))
-    return true
-  }
-}
-
-const setupPage = (store, id, query, action) => {
-  const { error, payload, cacheHit } = action
-  const { dispatch } = store
-  if (error) return
-  const state = store.getState()
-  const { currentCommunityId, posts, people } = state
-  const post = posts[id]
-  if (!post) return
-
-  const currentUser = people.current
-  let community_ids
-  if (post.parent_post_id) {
-    community_ids = posts[post.parent_post_id].community_ids
-  } else {
-    community_ids = post.community_ids
-  }
-  const communityId = includes(community_ids, currentCommunityId)
-    ? currentCommunityId
-    : (get('0', community_ids) || 'all')
-
-  if (payload && !payload.api) {
-    const { name, description, media } = payload
-    dispatch(setMetaTags(ogMetaTags(name, description, get('0', media))))
-  }
-
-  return Promise.all([
-    communityId !== currentCommunityId &&
-      saveCurrentCommunityId(dispatch, communityId, !!currentUser),
-
-    // Always load the full record for the current Community record
-    dispatch(fetchCommunity(communityId)),
-
-    // when this page is clicked into from a post list, fetchPost will cause a
-    // cache hit; however, there may be more comments than the 3 that were
-    // included in the list, so we have to call fetchComments to retrieve the
-    // rest. but when fetchPost did not cause a cache hit, we know that its
-    // response contained all comments, so we can skip the additional call.
-    cacheHit && post.numComments > 3 &&
-      dispatch(fetchComments(id, {refresh: true})).then(scroll),
-
-    // if this is a project, fetch the first page of results for child posts.
-    post.type === 'project' && dispatch(fetch(subject, id, {...query})),
-
-    get('action', query) === 'unfollow' && currentUser &&
-      dispatch(unfollowPost(post.id, currentUser.id)).then(({ error }) => !error &&
-        dispatch(notify('Notifications for this post are now off.',
-          {type: 'info', maxage: null})))
-  ])
-}
-
-const scroll = () => {
-  if (typeof window === 'undefined') return
-  let id = get(window.location.hash.match(/#comment-(\d+)$/), '1')
-  if (id) scrollToComment(id)
-}
