+++ src/routes/makeRoutes.js
@@ -23,7 +23,7 @@ import SkillsPrompt from '../containers/onboarding/SkillsPrompt'
 import WelcomePage from '../containers/onboarding/WelcomePage'
 import PersonProfile from '../containers/person/PersonProfile'
 import UserSettings from '../containers/user/UserSettings'
-import SinglePost from '../containers/SinglePost'
+import SinglePost from '../components/SinglePost'
 import NetworkProfile from '../containers/network/NetworkProfile'
 import NetworkPosts from '../containers/network/NetworkPosts'
 import NetworkMembers from '../containers/network/NetworkMembers'
